Inclusion of explore folder, throwaway scripts, restructured to accomodate
