+++ README.md
@@ -3,3 +3,8 @@ This is the repository containing my university final year project.
 The aim of this project is to perform a comparative, and evaluative study on current sentiment analysis methods. This involves
 comparing four tools: TextBlob, VADER, BERT and GPT-2. The study will weigh up the accuracy benefits against the computational
 costs of these models.
+
+Project Structure Overview:
+ - EXPLORE_REQUIREMENTS: Initial exploring and testing out APIs/Libraries manually to see what will need to go into the unit tests
+ - TESTING: The unit tests before development
+ - src: What is developed and used as part of the projects outputs
