+++ src/sa_models/sentiment_result.py
